server/requirements.txt
uvicorn
fastapi
sse-starlette
python-socketio
openai